# stewardess_openapi3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stewardess MCP Service",
    "description": "Local source-code repository tools for AI agents. Provides file reading, search, editing, git, and command execution.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://office:55702"
    }
  ],
  "paths": {
...
    "/api/files/read": {
      "get": {
        "operationId": "File_ReadFile",
        "summary": "Reads the full (or partial) content of a file.",
        "description": "Reads the full (or partial) content of a file.",
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxBytes",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "returnBase64",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/range": {
      "get": {
        "operationId": "File_ReadRange",
        "summary": "Reads a contiguous range of lines from a file.",
        "description": "Reads a contiguous range of lines from a file.",
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startLine",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "endLine",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "includeLineNumbers",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/read-multiple": {
      "post": {
        "operationId": "File_ReadMultiple",
        "summary": "Reads several files in a single POST call.",
        "description": "Reads several files in a single POST call.",
        "tags": [
          "File"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadMultipleFilesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/structure": {
      "get": {
        "operationId": "File_GetStructure",
        "summary": "Returns a best-effort structural summary of a code file.",
        "description": "Returns a best-effort structural summary of a code file.",
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/git/status": {
      "get": {
        "operationId": "Git_GetStatus",
        "summary": "Returns the current git status of the repository.\r\nOptionally restrict to a sub-path via ?path=src/MyLib.",
        "description": "Returns the current git status of the repository.\r\nOptionally restrict to a sub-path via ?path=src/MyLib.",
        "tags": [
          "Git"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/git/diff": {
      "post": {
        "operationId": "Git_GetDiff",
        "summary": "Returns the unified diff for working-tree or staged changes.",
        "description": "Returns the unified diff for working-tree or staged changes.",
        "tags": [
          "Git"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitDiffRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/git/diff/file": {
      "get": {
        "operationId": "Git_GetFileDiff",
        "summary": "Returns the unified diff for a single file.\r\nQuery params: path (required), scope (default: \"unstaged\").",
        "description": "Returns the unified diff for a single file.\r\nQuery params: path (required), scope (default: \"unstaged\").",
        "tags": [
          "Git"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/git/commits/{sha}": {
      "get": {
        "operationId": "Git_GetCommit",
        "summary": "Returns the full details of a single commit, including metadata, changed files,\r\nand optionally the diff patch.\r\nQuery params: includeDiff (default: true).",
        "description": "Returns the full details of a single commit, including metadata, changed files,\r\nand optionally the diff patch.\r\nQuery params: includeDiff (default: true).",
        "tags": [
          "Git"
        ],
        "parameters": [
          {
            "name": "sha",
            "in": "path",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeDiff",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/git/log": {
      "post": {
        "operationId": "Git_GetLog",
        "summary": "Returns the commit history for the repository or a sub-path.",
        "description": "Returns the commit history for the repository or a sub-path.",
        "tags": [
          "Git"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitLogRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReadMultipleFilesRequest": {
        "description": "Request to read multiple files in a single call.",
        "type": "object",
        "properties": {
          "paths": {
            "description": "List of relative paths to read.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxBytesPerFile": {
            "format": "int64",
            "description": "Per-file byte limit; null uses the global MaxFileReadBytes setting.",
            "type": "integer"
          }
        }
      },
      "GitDiffRequest": {
        "description": "Request to retrieve a git diff.",
        "type": "object",
        "properties": {
          "path": {
            "description": "Restrict diff to this file or directory.  Empty = whole repo.",
            "type": "string"
          },
          "scope": {
            "description": "Diff scope: \"unstaged\" (default), \"staged\", \"head\", \"commit:sha1..sha2\".",
            "type": "string"
          },
          "contextLines": {
            "format": "int32",
            "description": "Number of context lines around each change (default 3).",
            "type": "integer"
          }
        }
      },
      "GitLogRequest": {
        "description": "Request to retrieve the commit log.",
        "type": "object",
        "properties": {
          "path": {
            "description": "Restrict log to commits that touch this path (empty = all).",
            "type": "string"
          },
          "maxCount": {
            "format": "int32",
            "description": "Maximum number of commits to return (default 20).",
            "type": "integer"
          },
          "ref": {
            "description": "Branch or commit ref to start from; empty = HEAD.",
            "type": "string"
          },
          "author": {
            "description": "Restrict to commits by this author name or email.",
            "type": "string"
          },
          "since": {
            "description": "ISO-8601 date; return only commits after this date.",
            "type": "string"
          },
          "until": {
            "description": "ISO-8601 date; return only commits before this date.",
            "type": "string"
          }
        }
      },
...
    }
  }
}